refactor: trip with and without days components
