+++ src/components/trip/days/days-list.js
@@ -1,4 +1,4 @@
-import AbstractTripComponent from '../abstract-trip.js';
+import AbstractTripComponent from '../../abstract-trip.js';
 
 const createDaysListTemplate = () => {
   return (
